backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
audeer==2.2.1
audiofile==1.5.1
audmath==1.4.1
bcrypt==4.2.1
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
decorator==4.4.2
distro==1.9.0
dnspython==2.7.0
ffmpeg==1.4
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
groq==0.15.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.5
langdetect==1.0.9
MarkupSafe==3.0.2
moviepy==1.0.3
numpy==2.2.2
opencv-python==4.11.0.86
packaging==24.2
pillow==10.4.0
proglog==0.1.10
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pydub==0.25.1
PyJWT==2.10.1
pymongo==4.10.1
python-dotenv==1.0.1
requests==2.32.3
scipy==1.15.1
six==1.17.0
sniffio==1.3.1
soundfile==0.13.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
